Re-purpose disemvowel code for command line
Added `file_disemvowel.c` as specified
by the lab writeup and began making it compatible
with our original disemvowel code. InputFile and
outputFile are initialized in `main` method. Buffers are
created in `disemvowel` method which also is attempting to
continuously write to and dismevowel the buffers
while the input file has additional characters.
`copyNonVowels` now takes the input buffer and copies
non-vowels to the output buffer. Two tests are passing.