Fix read/write loop
We changed the order of how the input
is read and written to the output file
in our read/write loop. Changed inputFile
and outputFile to just be stdin and stdout
for now. Fixed our fread and fwrite calls
to use the size of chars. Four tests
passing, two are skipped.